change mailer to auto_post_fb and using pyramid_mailer in its
